+++ src/lib.rs
@@ -438,35 +438,10 @@ impl ParallelParam {
 
 macro_rules! sbs_encode_checks {
     (
-        nosep => $self:ident, $pieces:ident
+        $self:ident
     ) => {{
         if $self.parity_ready() {
             return Err(SBSError::TooManyCalls);
-        }
-        if $pieces.len() < $self.codec.total_shard_count() {
-            return Err(SBSError::RSError(Error::TooFewShards));
-        }
-        if $pieces.len() > $self.codec.total_shard_count() {
-            return Err(SBSError::RSError(Error::TooManyShards));
-        }
-    }};
-    (
-        sep => $self:ident, $data:ident, $parity:ident
-    ) => {{
-        if $self.parity_ready() {
-            return Err(SBSError::TooManyCalls);
-        }
-        if $data.len() < $self.codec.data_shard_count() {
-            return Err(SBSError::RSError(Error::TooFewDataShards));
-        }
-        if $data.len() > $self.codec.data_shard_count() {
-            return Err(SBSError::RSError(Error::TooManyDataShards));
-        }
-        if $parity.len() < $self.codec.parity_shard_count() {
-            return Err(SBSError::RSError(Error::TooFewParityShards));
-        }
-        if $parity.len() > $self.codec.parity_shard_count() {
-            return Err(SBSError::RSError(Error::TooManyParityShards));
         }
     }}
 }
@@ -534,7 +509,7 @@ impl<'a> ShardByShard<'a> {
     pub fn encode(&mut self,
                   slices : &mut [&mut [u8]])
                   -> Result<(), SBSError> {
-        sbs_encode_checks!(nosep => self, slices);
+        sbs_encode_checks!(self);
 
         let res = self.codec.encode_single(self.cur_input,
                                            slices);
@@ -550,7 +525,7 @@ impl<'a> ShardByShard<'a> {
                       data   : &[&[u8]],
                       parity : &mut [&mut[u8]])
                       -> Result<(), SBSError> {
-        sbs_encode_checks!(sep => self, data, parity);
+        sbs_encode_checks!(self);
 
         let res = self.codec.encode_single_sep(self.cur_input,
                                                &data[self.cur_input],
@@ -566,7 +541,7 @@ impl<'a> ShardByShard<'a> {
     pub fn encode_shard(&mut self,
                         shards : &mut [Shard])
                         -> Result<(), SBSError> {
-        sbs_encode_checks!(nosep => self, shards);
+        sbs_encode_checks!(self);
 
         let res = self.codec.encode_single_shard(self.cur_input,
                                                  shards);
@@ -582,7 +557,7 @@ impl<'a> ShardByShard<'a> {
                             data   : &[Shard],
                             parity : &mut [Shard])
                             -> Result<(), SBSError> {
-        sbs_encode_checks!(sep => self, data, parity);
+        sbs_encode_checks!(self);
 
         let res = self.codec.encode_single_shard_sep(self.cur_input,
                                                      &data[self.cur_input],