+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Improved shard utils
   - Added code to remove leftover parity shards in `reconstruct_data_shards`
     - This means one fewer gotcha of using the methods
+  - `ShardByShard` code clean up
 
 ## 2.4.0
   - Added more flexibility for `convert_2D_slices` macro